requirements.txt
matplotlib==3.6.2
numpy==1.23.4
Pillow==10.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
